+++ include/config.hpp
@@ -22,7 +22,7 @@ public:
 
     static int imageHeight(){return getInstance().height;}
 
-    static float cameraFps(){return getInstance().fps;}
+    static double cameraFps(){return getInstance().fps;}
 
     static int initMinCorners(){return getInstance().init_min_corners;}
 
@@ -32,13 +32,13 @@ public:
 
     static int initMinInliers(){return getInstance().init_min_inliers;}
 
-    static float initSigma(){return getInstance().init_sigma;}
+    static double initSigma(){return getInstance().init_sigma;}
 
-    static float initSigma2(){return getInstance().init_sigma2;}
+    static double initSigma2(){return getInstance().init_sigma2;}
 
-    static float initUnSigma(){return getInstance().init_unsigma;}
+    static double initUnSigma(){return getInstance().init_unsigma;}
 
-    static float initUnSigma2(){return getInstance().init_unsigma2;}
+    static double initUnSigma2(){return getInstance().init_unsigma2;}
 
     static int initMaxRansacIters(){return getInstance().init_max_iters;}
 
@@ -52,7 +52,7 @@ public:
 
     static int fastMinThreshold(){return getInstance().fast_min_threshold;}
 
-    static float fastMinEigen(){return getInstance().fast_min_eigen;}
+    static double fastMinEigen(){return getInstance().fast_min_eigen;}
 
 private:
     static Config& getInstance()
@@ -71,41 +71,41 @@ private:
         }
 
         //! camera parameters
-        fx = (float)fs["Camera.fx"];
-        fy = (float)fs["Camera.fy"];
-        cx = (float)fs["Camera.cx"];
-        cy = (float)fs["Camera.cy"];
+        fx = (double)fs["Camera.fx"];
+        fy = (double)fs["Camera.fy"];
+        cx = (double)fs["Camera.cx"];
+        cy = (double)fs["Camera.cy"];
 
-        k1 = (float)fs["Camera.k1"];
-        k2 = (float)fs["Camera.k2"];
-        p1 = (float)fs["Camera.p1"];
-        p2 = (float)fs["Camera.p2"];
+        k1 = (double)fs["Camera.k1"];
+        k2 = (double)fs["Camera.k2"];
+        p1 = (double)fs["Camera.p1"];
+        p2 = (double)fs["Camera.p2"];
 
-        K = cv::Mat::eye(3,3,CV_32F);
-        K.at<float>(0,0) = fx;
-        K.at<float>(1,1) = fy;
-        K.at<float>(0,2) = cx;
-        K.at<float>(1,2) = cy;
+        K = cv::Mat::eye(3,3,CV_64F);
+        K.at<double>(0,0) = fx;
+        K.at<double>(1,1) = fy;
+        K.at<double>(0,2) = cx;
+        K.at<double>(1,2) = cy;
 
-        DistCoef = cv::Mat::zeros(4,1,CV_32F);
-        DistCoef.at<float>(0) = k1;
-        DistCoef.at<float>(1) = k2;
-        DistCoef.at<float>(2) = p1;
-        DistCoef.at<float>(3) = p2;
+        DistCoef = cv::Mat::zeros(4,1,CV_64F);
+        DistCoef.at<double>(0) = k1;
+        DistCoef.at<double>(1) = k2;
+        DistCoef.at<double>(2) = p1;
+        DistCoef.at<double>(3) = p2;
 
         width = (int)fs["Camera.width"];
         height = (int)fs["Camera.height"];
-        fps = (float)fs["Camera.fps"];
+        fps = (double)fs["Camera.fps"];
 
         //! initializer parameters
         init_min_corners = (int)fs["Initializer.min_corners"];
         init_min_tracked = (int)fs["Initializer.min_tracked"];
         init_min_disparity = (int)fs["Initializer.min_disparity"];
         init_min_inliers = (int)fs["Initializer.min_inliers"];
-        init_sigma = (float)fs["Initializer.sigma"];
+        init_sigma = (double)fs["Initializer.sigma"];
         init_sigma2 = init_sigma*init_sigma;
         init_unsigma2 = init_sigma2 * 2.0 / (fx*fx+fy*fy);
-        init_unsigma = sqrtf(init_unsigma2);
+        init_unsigma = sqrt(init_unsigma2);
         init_max_iters = (int)fs["Initializer.ransac_max_iters"];
 
         //! FAST detector parameters
@@ -114,7 +114,7 @@ private:
         grid_max_fts = (int)fs["FastDetector.grid_max_fts"];
         fast_max_threshold = (int)fs["FastDetector.fast_max_threshold"];
         fast_min_threshold = (int)fs["FastDetector.fast_min_threshold"];
-        fast_min_eigen = (float)fs["FastDetector.fast_min_eigen"];
+        fast_min_eigen = (double)fs["FastDetector.fast_min_eigen"];
 
         fs.release();
     }
@@ -125,23 +125,23 @@ public:
 
 private:
     //! camera parameters
-    float fx, fy, cx, cy;
-    float k1, k2, p1, p2;
+    double fx, fy, cx, cy;
+    double k1, k2, p1, p2;
     cv::Mat K;
     cv::Mat DistCoef;
     int width;
     int height;
-    float fps;
+    double fps;
 
     //! initializer parameters
     int init_min_corners;
     int init_min_tracked;
     int init_min_disparity;
     int init_min_inliers;
-    float init_sigma;
-    float init_sigma2;
-    float init_unsigma;
-    float init_unsigma2;
+    double init_sigma;
+    double init_sigma2;
+    double init_unsigma;
+    double init_unsigma2;
     int init_max_iters;
 
     //! FAST detector parameters
@@ -150,7 +150,7 @@ private:
     int grid_max_fts;
     int fast_max_threshold;
     int fast_min_threshold;
-    float fast_min_eigen;
+    double fast_min_eigen;
 
 };
